Moving to the new project. (Work in progress)

+++ NewProject/Problem_026_050/Problem040.cpp
@@ -5,27 +5,27 @@ using namespace std;
 
 int64_t Problem40()
 {
-    const array<int, 7> Power    = { 1, 10, 100, 1000, 10000, 100000, 1000000 };
-    const array<int, 7> Position = { 1, 10, 100, 1000, 10000, 100000, 1000000 };
+    const array<int32_t, 7> Power    = { 1, 10, 100, 1000, 10000, 100000, 1000000 };
+    const array<int32_t, 7> Position = { 1, 10, 100, 1000, 10000, 100000, 1000000 };
 
-    int64_t answer = 1;
+    int32_t answer = 1;
 
-    for (const int pos : Position)
+    for (const int32_t pos : Position)
     {
-        int p = pos - 1;
-        int i;
+        int32_t p = pos - 1;
+        int32_t i;
         for (i = 1;; ++i)
         {
-            const int t = (Power[i] - Power[i - 1]) * i;
+            const int32_t t = (Power[i] - Power[i - 1]) * i;
             if (p < t)
                 break;
 
             p -= t;
         }
 
-        const int d = p / i;
-        const int r = p % i;
-        const int n = ((Power[i - 1] + d) / Power[i - r - 1]) % 10;
+        const int32_t d = p / i;
+        const int32_t r = p % i;
+        const int32_t n = ((Power[i - 1] + d) / Power[i - r - 1]) % 10;
 
         answer *= n;
     }
